+++ src/game/nlcore/game.ts
@@ -8,6 +8,8 @@ import { GuardWarningType } from "@player/guard";
 import { DefaultElements } from "../player/elements/elements";
 import { Plugins, IGamePluginRegistry } from "./game/plugin/plugin";
 import { LayoutRouter } from "../player/lib/PageRouter/router";
+import { KeyMap } from "./game/keyMap";
+import { KeyBindingType } from "./game/types";
 enum GameSettingsNamespace {
     game = "game",
 }
@@ -69,15 +71,7 @@ export class Game {
     static DefaultConfig: GameConfig = {
         app: {
             debug: false,
-            logger: {
-                log: false,
-                info: false,
-                warn: true,
-                error: true,
-                debug: false,
-                trace: false,
-                verbose: false,
-            },
+            logger: { log: false, info: false, warn: true, error: true, debug: false, trace: false, verbose: false, },
             inspector: false,
             guard: {
                 [GuardWarningType.invalidExposedStateUnmounting]: true,
@@ -90,7 +84,6 @@ export class Game {
         minHeight: 450,
         width: 1920,
         height: 1080,
-        skipKey: ["Control"],
         useWindowListener: true,
         ratioUpdateInterval: 50,
         preloadDelay: 100,
@@ -105,7 +98,6 @@ export class Game {
         showOverflow: false,
         maxRouterHistory: 10,
         screenshotQuality: 1,
-        nextKey: [" "],
         useAspectScale: true,
         autoForwardDelay: 3 * 1000,
         autoForwardDefaultPause: 1000,
@@ -150,6 +142,13 @@ export class Game {
      * Game settings
      */
     public preference: Preference<GamePreference> = new Preference<GamePreference>(Game.DefaultPreference);
+    /**
+     * Game key bindings
+     */
+    public keyMap: KeyMap = new KeyMap({
+        [KeyBindingType.skipAction]: ["Control"],
+        [KeyBindingType.nextAction]: [" "],
+    });
     /**
      * Plugin registry
      */